backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
GeoAlchemy2
PyJWT
passlib[bcrypt]
pydantic
pydantic-settings
httpx
email-validator
python-multipart
shapely